CHG: moved helper function to api
